requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
Django==4.0.4
django-filter==21.1
django-registration==3.2
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
